Don't remove footways that can be biked on
Some ways, such as https://www.openstreetmap.org/way/987797955 (#1) have
highway=footway, footway=crossing, bicycle=yes, and make up important
segments of bike routes. Currently, though, we strip these out as we
consider them sidewalks.

To fix that, incorporate the bikable footway to cycleway conversion
that's already in upstream[1] and then move our sidewalk removal logic
to after that conversion happens. This way, we'll still remove sidewalks
that aren't bikable, but the ones that are won't be footways by the time
we try to remove them.

[1]: https://github.com/openstreetmap/mkgmap/blob/453ed3a706a6dd0f05d1b78d89b688d59cad997e/resources/styles/default/lines#L61-L62